db: introduce SpanPolicy and SpanPolicyFunc
This change introduces the concept of a "span policy". Flushes and
compactions take into account the span policy when writing out ssts.

We replace the `RequiredInPlaceValueBound` option with a span policy.
The effective change is that we now split output ssts at the
boundaries where the policy applies, and the policy affects the writer
options for the entire sstable.

In the future, the span policy can include more things like hints to
optimize for size or tiering metadata.

For now, CockroachDB will define a `SpanPolicyFunc` with a single
static span. The intention is that in the future, span policies will
be tied to KV span configs.